add a bunch of tests and re-arrage file structure
